move verify method into extension method and expose publicly
